refactor: Extract option for re-use
