+++ src/python/pants/backend/python/dependency_inference/rules_test.py
@@ -35,7 +35,6 @@ from pants.backend.python.dependency_inference.rules import (
 from pants.backend.python.dependency_inference.subsystem import (
     InitFilesInference,
     PythonInferSubsystem,
-    UnownedDependencyUsage,
 )
 from pants.backend.python.macros import python_requirements
 from pants.backend.python.macros.python_requirements import PythonRequirementsTargetGenerator
@@ -49,6 +48,7 @@ from pants.backend.python.target_types import (
 from pants.backend.python.util_rules import ancestor_files
 from pants.core.target_types import FilesGeneratorTarget, ResourcesGeneratorTarget
 from pants.core.target_types import rules as core_target_types_rules
+from pants.core.util_rules.unowned_dependency_behavior import UnownedDependencyUsage
 from pants.engine.addresses import Address
 from pants.engine.internals.parametrize import Parametrize
 from pants.engine.rules import rule
