+++ src/python/pants/backend/python/dependency_inference/subsystem.py
@@ -5,18 +5,14 @@ from __future__ import annotations
 
 from enum import Enum
 
+from pants.core.util_rules.unowned_dependency_behavior import (
+    UnownedDependencyUsage,
+    UnownedDependencyUsageOption,
+)
 from pants.option.option_types import BoolOption, EnumOption, IntOption, StrListOption
 from pants.option.subsystem import Subsystem
 from pants.util.docutil import bin_name
 from pants.util.strutil import softwrap
-
-
-class UnownedDependencyUsage(Enum):
-    """What action to take when an inferred dependency is unowned."""
-
-    RaiseError = "error"
-    LogWarning = "warning"
-    DoNothing = "ignore"
 
 
 class InitFilesInference(Enum):
@@ -134,21 +130,12 @@ class PythonInferSubsystem(Subsystem):
             """
         ),
     )
-    unowned_dependency_behavior = EnumOption(
-        default=UnownedDependencyUsage.LogWarning,
-        help=softwrap(
-            """
-            How to handle imports that don't have an inferrable owner.
-
-            Usually when an import cannot be inferred, it represents an issue like Pants not being
-            properly configured, e.g. targets not set up. Often, missing dependencies will result
-            in confusing runtime errors like `ModuleNotFoundError`, so this option can be helpful
-            to error more eagerly.
-
-            To ignore any false positives, either add `# pants: no-infer-dep` to the line of the
+    unowned_dependency_behavior = UnownedDependencyUsageOption(
+        example_runtime_issue="`ModuleNotFoundError`",
+        how_to_ignore="""
+            either add `# pants: no-infer-dep` to the line of the
             import or put the import inside a `try: except ImportError:` block.
-            """
-        ),
+            """,
     )
     ambiguity_resolution = EnumOption(
         default=AmbiguityResolution.none,
